Rename rppi_validate.hpp to rppt_validate.hpp
