+++ .Doxyfile
@@ -949,17 +949,6 @@ INPUT                  = README.md \
                         include/rppt.h \
                         include/rppversion.h \
                         include/rppdefs.h \
-                        include/rppi_advanced_augmentations.h \
-                        include/rppi_arithmetic_operations.h \
-                        include/rppi_color_model_conversions.h \
-                        include/rppi_computer_vision.h \
-                        include/rppi_filter_operations.h \
-                        include/rppi_fused_functions.h \
-                        include/rppi_geometry_transforms.h \
-                        include/rppi_image_augmentations.h \
-                        include/rppi_logical_operations.h \
-                        include/rppi_morphological_transforms.h \
-                        include/rppi_statistical_operations.h \
                         include/rppt_tensor_arithmetic_operations.h \
                         include/rppt_tensor_audio_augmentations.h \
                         include/rppt_tensor_color_augmentations.h \
